refactor utils structure
