+++ src/utils/plugin.rs
@@ -1,181 +1,12 @@
-use flate2::read::GzDecoder;
-use pretty_duration::pretty_duration;
-use reqwest::{self, blocking::Response, header::HeaderValue};
-use serde_json::{self, Map, Value};
+use serde_json::Value;
+use tabled::{builder::Builder, settings::Style};
+
+use super::common::{Plugin, PluginPermission, ProjectFile};
 use std::{
-    env::var_os,
     fs::{self, File},
     io::{Read, Write},
     path::PathBuf,
-    time::{Duration, SystemTime, UNIX_EPOCH},
 };
-use tabled::{builder::Builder, settings::Style};
-use tar::Archive;
-
-#[derive(Debug)]
-pub struct ProjectPath {
-    path_components: Vec<String>,
-    is_dir: bool,
-}
-
-#[derive(Debug)]
-pub struct ProjectFile {
-    path: ProjectPath,
-    content: Vec<u8>,
-}
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub struct PluginPermission {
-    type_: String,
-}
-
-impl PluginPermission {
-    pub fn from_map(map: &Map<String, Value>) -> Result<PluginPermission, ()> {
-        let type_val: String = map
-            .get("type")
-            .map(|v| serde_json::from_value(v.clone()).or(Err(())))
-            .unwrap_or(Err(()))?;
-
-        Ok(PluginPermission { type_: type_val })
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Plugin {
-    pub folder: String,
-    pub name: Option<String>,
-    pub description: Option<String>,
-    pub homepage: Option<String>,
-    pub authors: Vec<String>,
-    pub tags: Vec<String>,
-    pub version: Option<String>,
-    pub licence: Option<String>,
-    pub permissions: Vec<PluginPermission>,
-}
-
-impl Plugin {
-    pub fn new() -> Self {
-        Plugin {
-            folder: String::new(),
-            name: None,
-            description: None,
-            homepage: None,
-            authors: Vec::new(),
-            tags: Vec::new(),
-            version: None,
-            licence: None,
-            permissions: Vec::new(),
-        }
-    }
-}
-
-pub fn get_files_from_gzip(buf: &Vec<u8>) -> Vec<ProjectFile> {
-    let dec = GzDecoder::new(&buf[..]);
-    let mut archive = Archive::new(dec);
-    let mut files = vec![];
-
-    for file in archive.entries().unwrap() {
-        let mut file = file.unwrap();
-
-        let full_path = file.path().unwrap();
-        let is_dir = file.header().entry_type().is_dir();
-
-        let mut full_path_components = full_path
-            .components()
-            .map(|comp| comp.as_os_str().to_string_lossy().into_owned())
-            .collect::<Vec<_>>();
-
-        full_path_components.remove(0);
-
-        if full_path_components.len() == 0 {
-            continue;
-        }
-
-        let project_path = ProjectPath {
-            is_dir: is_dir,
-            path_components: full_path_components,
-        };
-
-        let mut file_content: Vec<u8> = Vec::new();
-        file.read_to_end(&mut file_content).unwrap();
-
-        files.push(ProjectFile {
-            path: project_path,
-            content: file_content,
-        });
-    }
-
-    files
-}
-
-pub fn get_default_chatterino_path() -> Result<PathBuf, String> {
-    let machine_kind = if cfg!(target_os = "linux") {
-        Some("linux")
-    } else if cfg!(target_os = "windows") {
-        Some("windows")
-    } else {
-        None
-    };
-
-    match machine_kind {
-        Some("windows") => {
-            let mut roaming_path_buf = var_os("APPDATA").map(PathBuf::from).ok_or(
-                "Could not read %APPDATA% environment variable. Please use --path instead.",
-            )?;
-            roaming_path_buf.push("Chatterino2");
-            Ok(roaming_path_buf.to_owned())
-        }
-        Some("linux") => {
-            let mut home_path_buf = var_os("HOME")
-                .map(PathBuf::from)
-                .ok_or("Could not read $HOME environment variable. Please use --path instead.")?;
-            home_path_buf.push(".local/share/chatterino");
-            Ok(home_path_buf.to_owned())
-        }
-        _ => Err(
-            "Unsupported OS, cannot locate Chatterino folder. Please use --path instead."
-                .to_string(),
-        ),
-    }
-}
-
-pub fn handle_github_rate_limit(response: &Response) -> Result<(), String> {
-    let status = response.status();
-    if status.as_u16() == 403 || status.as_u16() == 429 {
-        // rate limit reached
-        let default_header_value = HeaderValue::from_str("").unwrap();
-        let reset_epoch = response
-            .headers()
-            .get("X-RateLimit-Reset")
-            .unwrap_or(&default_header_value)
-            .to_str()
-            .unwrap_or("")
-            .parse::<i32>()
-            .unwrap_or(-1);
-        let current_epoch = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_secs() as i32;
-
-        let mut error_str = "GitHub API rate limit reached!".to_owned();
-        let duration_str = pretty_duration(
-            &Duration::from_secs((reset_epoch - current_epoch) as u64),
-            None,
-        );
-        if reset_epoch != -1 {
-            error_str.push_str(&format!(" Resets in {}", duration_str));
-        }
-
-        return Err(error_str);
-    } else if !status.is_success() {
-        let status_str = status.as_str();
-        return Err(format!(
-            "GitHub API returned an unexpected status code: {status_str}"
-        ));
-    }
-
-    Ok(())
-}
 
 pub fn write_plugin_data(
     base_path: PathBuf,
